feat: add support for PSR-17
If the PSR-17 `ResponseFactoryInterface` is present in the container, it will now be used as the response factory for the `BasicAccess` instance.
Additionally, callable `ResponseInterface` services will now be wrapped in the (immediately deprecated) `CallableResponseFactoryDecorator`, which implements PSR-17.

To address tests that were working with the response factory, I ended up removing Prophecy usage in favor of native PHPUnit mocks.
I also discovered several prophecy declarations that were invalid (as they were never called) in the `BasicAccess` tests.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>